tests/wpt/tests/html/browsers/offline/browser-state/navigator_online_online.https.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>Offline Application Cache - navigator_online_online</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>

    <script>
      test(function() {
        assert_true(navigator.onLine, "onLine test");
      });
    </script>
  </body>
</html>
